Add First High-Level Standard Ruleset JSON Test
Fixes #3

This creates a basic framework for supporting high level tests where the
input and output are JSON files.

This should allow easier prototyping and building of new game types, as
they can be test driven easily.
The other benefit (and I admit) my primary one, is for sharing the
eventual test suite with other implementations/snakes!

Having the files be JSON is specifically interesting for the second use
case

As for the implementation it's pretty bare bones. It looks in a certain
directory for each file that matches `*.in.json`. This file is a JSON of
a BoardState and an array of SnakeMoves.
We run the standard ruleset using this input and compare it to the
matching `*.out.json` file from the same directory

This is also my first foray into golang, so any and all feedback is
appreciated!