Change SnakeResponse.Latency to be a string.
Currently, the CLI code has Latency as a uint32, which makes good
sense. However, both the [API
reference](https://docs.battlesnake.com/references/api#battlesnake)
and the production implementation of the battesnake server use a
decimal integer string to represent Latency instead.

It would be best if both implementations could be switched to use a
numerical value, but changing that in prod would be
backwards-incompatible and may break a bunch of people. It's better
for the behaviors to be consistent, and we can probably change the
simulator without bothering too many folks, so here we are.