Add PreUpdateBoard to the Map Interface
One of the issues that we've seen is that its harder to design maps
since you only can make changes AFTER the normal game logic is applied.

Here we add a method to the interface (and a map to test it) that allows
maps to hook into the editor before the moves are applied.

This is useful if you want to spawn a hazard or food, and want to see it
take effect in the same turn

Co-authored-by: JonathanArns <jonathan.arns@googlemail.com>